+++ lib/finitio/json_schema/proxy_type.rb
@@ -2,11 +2,10 @@ module Finitio
   class ProxyType
 
     def to_json_schema(*args, &bl)
-      if @target
-        @target.to_json_schema(*args, &bl)
-      else
-        "object"
-      end
+      # ProxyType is supposed to be used only for recursive types.
+      # We don't have support for references yet, so let just
+      # generate an object here in the mean time.
+      { type: "object" }
     end
 
   end # module ProxyType