+++ lib/finitio/syntax.rb
@@ -21,18 +21,18 @@ module Finitio
     end
 
     def self.ast(source)
-      parse(source, root: "system").to_ast
+      parse(source, root: 'system').to_ast
     end
 
     def self.compile(source, cpl = nil)
       cpl = Compilation.coerce(cpl, source)
-      parse(source, root: "system").compile(cpl)
-      cpl.resolve_proxies
+      parse(source, root: 'system').compile(cpl)
+      ProxyResolver.new.resolve!(cpl.system)
     end
 
     def self.compile_type(source, cpl = nil)
       cpl = Compilation.coerce(cpl, source)
-      parse(source, root: "type").compile(cpl)
+      parse(source, root: 'type').compile(cpl)
     end
 
   end # module Syntax